requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.2.0
bcc==0.18.0
beautifulsoup4==4.11.1
beniget==0.4.1
binwalk==2.3.3
black==22.6.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
CacheControl==0.12.11
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.1.0
chrome-gnome-shell==0.0.0
cleo==2.0.1
click==8.1.3
click-help-colors==0.9.1
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
crashtest==0.4.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dateparser==1.1.8
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
dulwich==0.20.50
fastapi==0.78.0
filelock==3.8.2
fonttools==4.29.1
frozenlist==1.4.0
fs==2.4.12
gast==0.5.2
gotify==0.5.0
gpg==1.16.0
h11==0.14.0
hashID==3.1.4
hidpidaemon==18.4.6
html5lib==1.1
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
idna==3.3
importlib-metadata==5.1.0
jaraco.classes==3.2.3
jeepney==0.7.1
jsonschema==4.17.3
kernelstub==3.1.4
keyring==23.11.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
matplotlib==3.5.1
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.4
multidict==6.0.3
mutagen==1.46.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.9.2
platformdirs==2.5.2
ply==3.11
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
pop-transition==1.1.2
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodome==3.19.0
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.10.2
pydbus==0.6.0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.19.2
python-apt==2.4.0+ubuntu2
python-binance==1.0.19
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-slugify==6.1.2
python-xlib==0.29
pythran==0.10.0
pytube==12.1.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
rapidfuzz==2.12.0
redis==4.3.4
regex==2023.8.8
repolib==2.2.1
repoman==1.4.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
schedule==1.2.1
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
shellingham==1.5.0
shtab==1.6.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
spotipy==2.20.0
starlette==0.19.1
sympy==1.9
systemd-python==234
talipp==1.9.1
termcolor==2.3.0
text-unidecode==1.3
tldr==3.2.0
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2022.12.1
typing_extensions==4.4.0
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.8.0
unicodedata2==14.0.0
urllib3==1.26.5
uvicorn==0.17.6
vboxapi==1.0
virtualenv==20.14.1
wadllib==1.3.6
webencodings==0.5.1
websockets==9.1
wrapt==1.14.1
xdg==5.1.1
xkit==0.0.0
yarl==1.8.2
yt-dlp==2023.6.22
ytmusicapi==0.22.0
zipp==1.0.0
